Move worktree resolution function from fs to git store
