+++ crates/project/tests/integration/git_store.rs
@@ -1535,3 +1535,85 @@ mod trust_tests {
         });
     }
 }
+
+mod resolve_worktree_tests {
+    use fs::FakeFs;
+    use gpui::TestAppContext;
+    use project::git_store::resolve_git_worktree_to_main_repo;
+    use serde_json::json;
+    use std::path::{Path, PathBuf};
+
+    #[gpui::test]
+    async fn test_resolve_git_worktree_to_main_repo(cx: &mut TestAppContext) {
+        let fs = FakeFs::new(cx.executor());
+        // Set up a main repo with a worktree entry
+        fs.insert_tree(
+            "/main-repo",
+            json!({
+                ".git": {
+                    "worktrees": {
+                        "feature": {
+                            "commondir": "../../",
+                            "HEAD": "ref: refs/heads/feature"
+                        }
+                    }
+                },
+                "src": { "main.rs": "" }
+            }),
+        )
+        .await;
+        // Set up a worktree checkout pointing back to the main repo
+        fs.insert_tree(
+            "/worktree-checkout",
+            json!({
+                ".git": "gitdir: /main-repo/.git/worktrees/feature",
+                "src": { "main.rs": "" }
+            }),
+        )
+        .await;
+
+        let result =
+            resolve_git_worktree_to_main_repo(fs.as_ref(), Path::new("/worktree-checkout")).await;
+        assert_eq!(result, Some(PathBuf::from("/main-repo")));
+    }
+
+    #[gpui::test]
+    async fn test_resolve_git_worktree_normal_repo_returns_none(cx: &mut TestAppContext) {
+        let fs = FakeFs::new(cx.executor());
+        fs.insert_tree(
+            "/repo",
+            json!({
+                ".git": {},
+                "src": { "main.rs": "" }
+            }),
+        )
+        .await;
+
+        let result = resolve_git_worktree_to_main_repo(fs.as_ref(), Path::new("/repo")).await;
+        assert_eq!(result, None);
+    }
+
+    #[gpui::test]
+    async fn test_resolve_git_worktree_no_git_returns_none(cx: &mut TestAppContext) {
+        let fs = FakeFs::new(cx.executor());
+        fs.insert_tree(
+            "/plain",
+            json!({
+                "src": { "main.rs": "" }
+            }),
+        )
+        .await;
+
+        let result = resolve_git_worktree_to_main_repo(fs.as_ref(), Path::new("/plain")).await;
+        assert_eq!(result, None);
+    }
+
+    #[gpui::test]
+    async fn test_resolve_git_worktree_nonexistent_returns_none(cx: &mut TestAppContext) {
+        let fs = FakeFs::new(cx.executor());
+
+        let result =
+            resolve_git_worktree_to_main_repo(fs.as_ref(), Path::new("/does-not-exist")).await;
+        assert_eq!(result, None);
+    }
+}
